pyproject.toml
[project]
name = "nonebot-plugin-auto-repeater"
version = "0.2.0"
description = "A plugin based on NoneBot2, auto repeater message in group."
readme = "README.md"
license = { text = "GPL" }
authors = [
    { name = "zhulinyv", email = "zhulinyv2005@outlook.com" },
    { name = "MaxCrazy1101", email = "alex02newton@gmail.com" },
]
requires-python = ">=3.10"
dependencies = [
    "nonebot-adapter-onebot>=2.4.6,<3.0.0",
    "nonebot-plugin-alconna>=0.54.2,<1.0.0",
    "nonebot-plugin-apscheduler>=0.5.0,<1.0.0",
    "nonebot-plugin-localstore>=0.7.4,<1.0.0",
    "nonebot2>=2.4.1,<3.0.0",
]

[project.urls]
homepage = "https://github.com/MaxCrazy1101/nonebot-plugin-auto-repeater"
repository = "https://github.com/MaxCrazy1101/nonebot-plugin-auto-repeater"

[project.optional-dependencies]
dev = ["nonebot2[fastapi]>=2.4.1"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.nonebot]
plugins = ["nonebot_plugin_auto_repeater"]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]

[tool.isort]
profile = "black"
line_length = 90
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 90
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901"]

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
typeCheckingMode = "basic"
